pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "1.0.0"
description = "Modern FastAPI project with authentication and authorization"
authors = ["Adilson <adilsonlimadepaula@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
alembic = "^1.12.0"
asyncpg = "^0.28.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.6"
email-validator = "^2.0.0.post2"
celery = { extras = ["redis"], version = "^5.3.4", optional = true }
redis = { version = "^5.0.0", optional = true }
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
black = "^23.9.1"
isort = "^5.12.0"
autoflake = "^2.2.1"
mypy = "^1.5.1"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"
bcrypt = "^3.2.2"
pylint = "^3.3.6"

[tool.poetry.extras]
worker = ["celery", "redis"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["jose.*", "passlib.*", "alembic.*", "celery.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"